requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-groq
pydantic
python-dotenv
google-genai
numpy
